backend/chatbot/requirements.txt
Flask==2.2.5
requests==2.26.0
flask-cors==3.0.10
python-dotenv==0.19.2
googlemaps==4.10.0
openai==1.3.0
httpx==0.24.1
pydantic==2.6.1
pydantic-core==2.16.2
flask-sqlalchemy==3.1.1